requirements.txt
streamlit==1.42.2
python-jobspy==1.1.77
pandas>=2.1.0
beautifulsoup4>=4.12.2
requests>=2.31.0
numpy==1.26.3
pydantic>=2.3.0
tls-client>=1.0.1
markdownify>=0.13.1
regex>=2024.4.28 